@@ -2,13 +2,14 @@ name: audioplayers_test_server
 publish_to: none
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
+  async: ^2.11.0
   shelf: ^1.2.0
   shelf_router: ^1.0.0
   shelf_static: ^1.0.0
 
 dev_dependencies:
-  flame_lint: ^0.2.0
-  http: ^0.13.0
+  flame_lint: ^1.0.0
+  http: '>=0.13.1 <2.0.0'
