@@ -1,6 +1,6 @@
 name: audioplayers
 description: A Flutter plugin to play multiple audio files simultaneously
-version: 4.0.1
+version: 5.1.0
 homepage: https://github.com/bluefireteam/audioplayers
 repository: https://github.com/bluefireteam/audioplayers/tree/master/packages/audioplayers
 
@@ -21,16 +21,16 @@ flutter:
         default_package: audioplayers_windows
 
 dependencies:
-  audioplayers_android: ^3.0.1
-  audioplayers_darwin: ^4.0.1
-  audioplayers_linux: ^2.0.1
-  audioplayers_platform_interface: ^5.0.0
-  audioplayers_web: ^3.0.1
-  audioplayers_windows: ^2.0.1
-  file: ^6.1.0
+  audioplayers_android: ^4.0.1
+  audioplayers_darwin: ^5.0.1
+  audioplayers_linux: ^3.0.0
+  audioplayers_platform_interface: ^6.0.0
+  audioplayers_web: ^4.0.0
+  audioplayers_windows: ^3.0.0
+  file: '>=6.1.0 <8.0.0'
   flutter:
     sdk: flutter
-  http: ^0.13.5
+  http: '>=0.13.1 <2.0.0'
   meta: ^1.7.0
   path_provider: ^2.0.12
   synchronized: ^3.0.0
@@ -38,11 +38,11 @@ dependencies:
 
 dev_dependencies:
   dartdoc: ^6.1.5
-  flame_lint: ^0.2.0
+  flame_lint: ^1.0.0
   flutter_test:
     sdk: flutter
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ">=3.3.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.10.0'
 
