@@ -1,6 +1,6 @@
 name: audioplayers_linux
 description: Linux implementation of audioplayers, a Flutter plugin to play multiple audio files simultaneously
-version: 2.0.1
+version: 3.0.0
 homepage: https://github.com/bluefireteam/audioplayers
 repository: https://github.com/bluefireteam/audioplayers/tree/master/packages/audioplayers_linux
 
@@ -12,17 +12,17 @@ flutter:
         pluginClass: AudioplayersLinuxPlugin
 
 dependencies:
-  audioplayers_platform_interface: ^5.0.0
+  audioplayers_platform_interface: ^6.0.0
   flutter:
     sdk: flutter
 
 dev_dependencies:
   dartdoc: ^6.1.5
-  flame_lint: ^0.2.0
+  flame_lint: ^1.0.0
   flutter_test:
     sdk: flutter
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ">=3.3.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.13.0'
 
