@@ -1,10 +1,11 @@
 name: audioplayers_platform_interface
 description: The platform interface for audioplayers, a Flutter plugin to play multiple audio files simultaneously
-version: 5.0.0
+version: 6.0.0
 homepage: https://github.com/bluefireteam/audioplayers
 repository: https://github.com/bluefireteam/audioplayers/tree/master/packages/audioplayers_platform_interface
 
 dependencies:
+  collection: ^1.17.1
   flutter:
     sdk: flutter
   meta: ^1.7.0
@@ -12,10 +13,10 @@ dependencies:
 
 dev_dependencies:
   dartdoc: ^6.1.5
-  flame_lint: ^0.2.0
+  flame_lint: ^1.0.0
   flutter_test:
     sdk: flutter
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ">=3.3.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.10.0'
