@@ -1,7 +1,7 @@
 name: audioplayers_workspace
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dev_dependencies:
   melos: ^3.0.0
