@@ -2,5 +2,4 @@ package xyz.luan.audioplayers.example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()
