@@ -3,7 +3,10 @@ import 'dart:io';
 
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
+import 'package:shelf_router/shelf_router.dart' as shelf_router;
 import 'package:shelf_static/shelf_static.dart' as shelf_static;
+
+import 'stream_route.dart';
 
 Future<void> main() async {
   final port = int.parse(Platform.environment['PORT'] ?? '8080');
@@ -12,7 +15,22 @@ Future<void> main() async {
   final isLogRequests =
       (Platform.environment['LOG_REQUESTS'] ?? 'false') == 'true';
 
-  final cascade = Cascade().add(_staticHandler);
+  final publicStaticHandler = shelf_static.createStaticHandler(
+    'public',
+    defaultDocument: 'index.html',
+    serveFilesOutsidePath: true,
+  );
+
+  final recordMode = bool.parse(Platform.environment['RECORD_MODE'] ?? 'false');
+  final liveMode =
+      recordMode || bool.parse(Platform.environment['LIVE_MODE'] ?? 'false');
+  final routeHandler = shelf_router.Router()
+    ..mount(
+      '/stream',
+      StreamRoute(isLiveMode: liveMode, isRecordMode: recordMode).pipeline,
+    );
+
+  final cascade = Cascade().add(publicStaticHandler).add(routeHandler);
 
   var pipeline = const Pipeline();
   if (isLogRequests) {
@@ -36,16 +54,7 @@ Future<void> main() async {
     port,
   );
 
-  // TODO(Gustl22): provide an audio streaming endpoint:
-  // Inspiration: https://github.com/daspinola/video-stream-sample/blob/master/server.js
-
   print(
     'Serving at http://${server.address.host}:${server.port} with latency of $requestTimeoutMillis ms',
   );
 }
-
-final _staticHandler = shelf_static.createStaticHandler(
-  'public',
-  defaultDocument: 'index.html',
-  serveFilesOutsidePath: true,
-);
