@@ -19,7 +19,9 @@ class FlutterWindow : public Win32Window {
   // Win32Window:
   bool OnCreate() override;
   void OnDestroy() override;
-  LRESULT MessageHandler(HWND window, UINT const message, WPARAM const wparam,
+  LRESULT MessageHandler(HWND window,
+                         UINT const message,
+                         WPARAM const wparam,
                          LPARAM const lparam) noexcept override;
 
  private:
