@@ -1,7 +1,7 @@
 package xyz.luan.audioplayers
 
-import org.junit.jupiter.api.Test
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
 
 internal class ToConstantCaseTest {
     @Test
