@@ -1,15 +1,15 @@
 extension MapParser on Map<dynamic, dynamic> {
   bool containsKey(String key) => this.containsKey(key);
 
-  String getString(String key) {
-    return this[key] as String;
+  String? getString(String key) {
+    return this[key] as String?;
   }
 
-  int getInt(String key) {
-    return this[key] as int;
+  int? getInt(String key) {
+    return this[key] as int?;
   }
 
-  bool getBool(String key) {
-    return this[key] as bool;
+  bool? getBool(String key) {
+    return this[key] as bool?;
   }
 }
